requirements.txt

hikari==2.0.0.dev120
hikari-lightbulb==2.3.3
requests==2.31.0
